Allow coverage to be run multiple times cleanly
Currently if this is run as a part of a watch task, the statement count will
carry over from the previous run which will throw off the coverage
percentage over time. The same applies to the ok status.

This will define both status and totals on a per run basis so each run is
clean, whether the task is run on watch or manually from the command line.